+++ backend/src/jv2backend/__main__.py
@@ -3,6 +3,12 @@
 
 # Don't assume that the command which called the package is the main package
 # This is needed for nix
-# import sys
-# import os.path
-# sys.argv[0] = os.path.dirname(sys.argv[0])
+import sys
+import os.path
+sys.argv[0] = os.path.dirname(sys.argv[0])
+
+import app
+
+
+if __name__ == "__main__":
+    app.go()
