+++ backend/src/jv2backend/classes/collection.py
