+++ backend/src/jv2backend/classes/journal.py
