+++ backend/src/jv2backend/routes/journal.py
